LPS-138138 Extract logic out to util class
